+++ src/main-web.ts
@@ -1,6 +1,5 @@
 import './app.css'
 import App from './App.svelte'
-// import App from './lib/mambo.svelte';
 
 const app = new App({
   target: document.getElementById('app'),
