+++ src/view/view.ts
@@ -1,6 +1,6 @@
 import { WebGLRenderer, Scene, Camera, PerspectiveCamera, LineSegments, Material } from "three";
 import { Points, Vector3, Matrix4 } from "three";
-import type { coordinates3D } from "../utils/geometry";
+import type { coordinates3D } from "../model/galaxy";
 import { OrbitControls } from "three/examples/jsm/controls/OrbitControls.js";
 import { pointGeometryFromData, materialFromData } from "./points.js";
 import { lineGeometryFromData } from "./lines.js";
